- Accelerometer controls take store initial x and y values as zero when GameFragment.resume() is called.  - High scores display properly.  - All Java import declarations reorganized.  - Shot and Player track hitboxes.  - Foreground and Background classes renamed to ForegroundView and BackgroundView.  - Player fires cannon when charge meter is full.  - Tapping screen fires EMP, and Shots are removed when within EMP radius.  - Tapping pause button pauses game.

+++ trunk/Untouchable/src/com/example/untouchable/MainActivity.java
@@ -47,15 +47,15 @@ public class MainActivity extends Activity {
     			break;*/
     			
     		case R.id.easy:
-    			((StartFragment)frag).setDifficulty(1);
+    			((StartFragment)frag).setDifficulty((short) 1);
     			break;
     			
     		case R.id.normal:
-    			((StartFragment)frag).setDifficulty(2);
+    			((StartFragment)frag).setDifficulty((short) 2);
     			break;
     			
     		case R.id.hard:
-    			((StartFragment)frag).setDifficulty(3);
+    			((StartFragment)frag).setDifficulty((short) 3);
     			break;
 			
     		case R.id.startButton:
